page-read/xfer: add ability to open image hierarchy for shmem
In order to restore deduplicated anonymous shared memory we need to open
it's parent pagemap images. Code that opens parent pagemap images already
exists for anonymous private memory. All we need to do is to remove couple
of checks from existing code. Also we need to rename pid to id because now
we can pass either pid or shmid and the actual meaning depends on pr_flags.

Signed-off-by: Fyodor Bocharov <fbocharov@yandex.ru>
Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>