mem: make page_in_parent routine extern
If we want to dedup anon shared memory we need to check for every page if
it's in parent -- this routine we already have, just need to make it extern.
But unfortunately for shmem we have only 1 bit per page so we have to change
routine signature but not logic because it operates only with one bit from
PME -- SOFT-DIRTY bit.

Signed-off-by: Fyodor Bocharov <fbocharov@yandex.ru>
Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>