shmem: implement dirty page tracking for anon shared memory
To track anon shared vma pages dirtiness we create a bitmap for it.
Each bit in this bitmap corresponds to particular page in vma. Bit with
value of 1 in this map states that page is dirty. Bit 0 states that the
page is clean.
Now we can collect shared vmas in __parasite_dump_pages_seized routine
instead of doing it from dump_task_mm because there we have pagemaps of
dumpee but pmc should be big enough to hold the longest of private and
shared vmas.

Signed-off-by: Fyodor Bocharov <fbocharov@yandex.ru>
Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>