shmem: add used pages tracking
To correctly dedup shared anon memory we need to know which pages was used
and dump only them -- this allows us to reduce size of first dump. To do
that we create bitmap for each shared vma where each bit corresponds to
particular page in vma. Bit with value 1 in this bitmap states that page
was used. Bit 0 states that page was not used.

Signed-off-by: Fyodor Bocharov <fbocharov@yandex.ru>
Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>